pyproject.toml
[build-system]
requires = ["setuptools>=82.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "langcore"
version = "0.1.0"
description = "LangCore: LLM-powered structured information extraction from text"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    {name = "Gabriel Ignat", email = "gabriel.ignat@hotmail.com"}
]
dependencies = [
    "absl-py>=2.4.0",
    "aiohttp>=3.13.3",
    "async_timeout>=4.0.0",
    "exceptiongroup>=1.3.0",
    "google-genai>=1.39.0",
    "google-cloud-storage>=2.14.0",
    "intervaltree>=3.2.1",
    "ml-collections>=1.1.0",
    "more-itertools>=10.8.0",
    "numpy>=2.4.2",
    "pandas>=1.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "PyYAML>=6.0.3",
    "rapidfuzz>=3.14.3",
    "regex>=2026.2.19",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
    "typing-extensions>=4.15.0"
]

[project.urls]
"Homepage" = "https://github.com/IgnatG/langcore"
"Repository" = "https://github.com/IgnatG/langcore"
"Documentation" = "https://github.com/IgnatG/langcore/blob/main/README.md"
"Bug Tracker" = "https://github.com/IgnatG/langcore/issues"
"Changelog" = "https://github.com/IgnatG/langcore/blob/main/CHANGELOG.md"
"Upstream" = "https://github.com/google/langextract"

[project.optional-dependencies]
openai = ["openai>=1.50.0"]
all = ["openai>=1.50.0"]
test = [
    "pytest>=9.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "ruff>=0.15.2",
]

[tool.setuptools]
packages = [
    "langcore",
    "langcore._compat",
    "langcore.core",
    "langcore.providers",
    "langcore.providers.schemas"
]
include-package-data = true

[tool.setuptools.package-data]
langcore = ["py.typed"]

# Provider discovery mechanism for built-in and third-party providers
[project.entry-points."langcore.providers"]
gemini = "langcore.providers.gemini:GeminiLanguageModel"
ollama = "langcore.providers.ollama:OllamaLanguageModel"
openai = "langcore.providers.openai:OpenAILanguageModel"

[tool.setuptools.exclude-package-data]
"*" = [
    "docs*",
    "tests*",
    "kokoro*",
    "*.gif",
    "*.svg",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "*_test.py"
python_classes = "Test*"
python_functions = "test_*"
# Show extra test summary info
addopts = "-ra"
markers = [
    "live_api: marks tests as requiring live API access",
    "requires_pip: marks tests that perform pip install/uninstall operations",
    "integration: marks integration tests that test multiple components together",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "TCH", "RUF"]
ignore = ["E501", "TC001", "TC002", "TC003", "B905", "RUF001"]

[tool.ruff.lint.per-file-ignores]
"benchmarks/*" = ["E741", "B007"]
"tests/*" = ["N802", "RUF012", "SIM117"]
"langcore/data.py" = ["F403"]
"langcore/tokenizer.py" = ["F403"]
"langcore/resolver.py" = ["N813", "B008"]
"langcore/annotation.py" = ["SIM102"]
"langcore/core/tokenizer.py" = ["SIM102"]
"langcore/core/schema.py" = ["B027"]
"langcore/hooks.py" = ["SIM105"]
"langcore/providers/ollama.py" = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["langcore"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["langcore/__init__.py:__version__"]
branch = "main"
build_command = "pip install build && python -m build"
commit_message = "chore(release): v{version}"
tag_format = "v{version}"

[tool.semantic_release.changelog]
generate = true

[tool.semantic_release.publish]
upload_to_pypi = true
